--- a/include/storage/io.hpp
+++ b/include/storage/io.hpp
@@ -18,6 +18,7 @@
 
 #include <cerrno>
 #include <cstring>
+#include <iostream>
 #include <tuple>
 #include <type_traits>