--- a/include/util/debug.hpp
+++ b/include/util/debug.hpp
@@ -12,7 +12,7 @@
 #include "util/typedefs.hpp"
 
 #include <iomanip>
-#include <iostream>
+#include <ostream>
 #include <sstream>
 #include <string>
 #include <vector>