--- a/include/util/exception_utils.hpp
+++ b/include/util/exception_utils.hpp
@@ -1,6 +1,7 @@
 #ifndef SOURCE_MACROS_HPP
 #define SOURCE_MACROS_HPP
 
+#include <cstring>
 #include <string>
 
 // Helper macros, don't use these ones