--- a/include/util/exception_utils.hpp
+++ b/include/util/exception_utils.hpp
@@ -1,5 +1,5 @@
-#ifndef SOURCE_MACROS_HPP
-#define SOURCE_MACROS_HPP
+#ifndef EXCEPTION_UTILS_HPP
+#define EXCEPTION_UTILS_HPP
 
 #include <cstring>
 #include <string>
@@ -14,4 +14,4 @@
 // This is the macro to use
 #define SOURCE_REF (OSRM_SOURCE_FILE_ + ":" + std::to_string(__LINE__))
 
-#endif // SOURCE_MACROS_HPP
+#endif // EXCEPTION_UTILS_HPP