--- a/include/util/lua_util.hpp
+++ b/include/util/lua_util.hpp
@@ -10,7 +10,6 @@ extern "C"
 
 #include <boost/filesystem.hpp>
 
-#include <iostream>
 #include <string>
 
 namespace osrm::util