--- a/include/util/node_based_graph.hpp
+++ b/include/util/node_based_graph.hpp
@@ -9,7 +9,6 @@
 
 #include <tbb/parallel_sort.h>
 
-#include <iostream>
 #include <memory>
 #include <utility>