--- a/src/extractor/compressed_edge_container.cpp
+++ b/src/extractor/compressed_edge_container.cpp
@@ -9,8 +9,6 @@
 #include <limits>
 #include <string>
 
-#include <iostream>
-
 namespace osrm::extractor
 {