--- a/src/extractor/extractor.cpp
+++ b/src/extractor/extractor.cpp
@@ -50,7 +50,6 @@
 #include <atomic>
 #include <bitset>
 #include <chrono>
-#include <iostream>
 #include <memory>
 #include <thread>
 #include <tuple>