--- a/src/storage/storage.cpp
+++ b/src/storage/storage.cpp
@@ -31,7 +31,6 @@
 #include <cstdint>
 
 #include <fstream>
-#include <iostream>
 #include <iterator>
 #include <new>
 #include <string>