--- a/src/tools/extract.cpp
+++ b/src/tools/extract.cpp
@@ -9,6 +9,7 @@
 
 #include <cstdlib>
 #include <exception>
+#include <iostream>
 #include <new>
 #include <thread>