--- a/src/util/exception.cpp
+++ b/src/util/exception.cpp
@@ -1,7 +1,5 @@
 #include "util/exception.hpp"
 
-#include <utility>
-
 // This function exists to 'anchor' the class, and stop the compiler from
 // copying vtable and RTTI info into every object file that includes
 // this header. (Caught by -Wweak-vtables under Clang.)