# Issue

including iostream is a code smell :grin: 

<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1149.5<br/>plain u32: 1152.92<br/>aliased double: 1192.73<br/>plain double: 1226.24 | aliased u32: 1140.04<br/>plain u32: 1135.96<br/>aliased double: 1191.97<br/>plain double: 1179.31 |
| json-render | String: 8.65366ms<br/>Stringstream: 11.9421ms<br/>Vector: 7.72664ms | String: 8.3811ms<br/>Stringstream: 11.5928ms<br/>Vector: 7.48348ms |
| match_ch | Default radius: <br/>4.45105ms/req at 82 coordinate<br/>0.0542811ms/coordinate<br/>Radius 5m: <br/>4.44301ms/req at 82 coordinate<br/>0.0541831ms/coordinate<br/>Radius 10m: <br/>15.1047ms/req at 82 coordinate<br/>0.184203ms/coordinate<br/>Radius 15m: <br/>36.8425ms/req at 82 coordinate<br/>0.449299ms/coordinate<br/>Radius 30m: <br/>313.75ms/req at 82 coordinate<br/>3.82622ms/coordinate | Default radius: <br/>4.45681ms/req at 82 coordinate<br/>0.0543513ms/coordinate<br/>Radius 5m: <br/>4.42715ms/req at 82 coordinate<br/>0.0539896ms/coordinate<br/>Radius 10m: <br/>15.0754ms/req at 82 coordinate<br/>0.183846ms/coordinate<br/>Radius 15m: <br/>36.7688ms/req at 82 coordinate<br/>0.4484ms/coordinate<br/>Radius 30m: <br/>313.635ms/req at 82 coordinate<br/>3.82482ms/coordinate |
| match_mld | Default radius: <br/>2.83922ms/req at 82 coordinate<br/>0.0346246ms/coordinate<br/>Radius 5m: <br/>3.46332ms/req at 82 coordinate<br/>0.0422356ms/coordinate<br/>Radius 10m: <br/>10.2364ms/req at 82 coordinate<br/>0.124835ms/coordinate<br/>Radius 15m: <br/>25.959ms/req at 82 coordinate<br/>0.316573ms/coordinate<br/>Radius 30m: <br/>302.505ms/req at 82 coordinate<br/>3.68909ms/coordinate | Default radius: <br/>2.82378ms/req at 82 coordinate<br/>0.0344363ms/coordinate<br/>Radius 5m: <br/>2.7982ms/req at 82 coordinate<br/>0.0341244ms/coordinate<br/>Radius 10m: <br/>10.1089ms/req at 82 coordinate<br/>0.123279ms/coordinate<br/>Radius 15m: <br/>25.6753ms/req at 82 coordinate<br/>0.313113ms/coordinate<br/>Radius 30m: <br/>298.037ms/req at 82 coordinate<br/>3.6346ms/coordinate |
| packedvector | random write:<br/>std::vector 11533.7 ms<br/>util::packed_vector 78341.3 ms<br/>slowdown: 6.79238<br/>random read:<br/>std::vector 11336.6 ms<br/>util::packed_vector 33666.9 ms<br/>slowdown: 2.96975 | random write:<br/>std::vector 11276.1 ms<br/>util::packed_vector 76592.5 ms<br/>slowdown: 6.79247<br/>random read:<br/>std::vector 11239.4 ms<br/>util::packed_vector 32473.2 ms<br/>slowdown: 2.88923 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>592.398ms<br/>0.592398ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>374.506ms<br/>0.374506ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>758.339ms<br/>0.758339ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>160.27ms<br/>0.16027ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>102.934ms<br/>0.102934ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>143.359ms<br/>0.143359ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>160.865ms<br/>0.160865ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>102.698ms<br/>0.102698ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>142.837ms<br/>0.142837ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>596.461ms<br/>0.596461ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>379.358ms<br/>0.379358ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>768.937ms<br/>0.768937ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>161.377ms<br/>0.161377ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>103.603ms<br/>0.103603ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>144.056ms<br/>0.144056ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>159.623ms<br/>0.159623ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>103.457ms<br/>0.103457ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>144.124ms<br/>0.144124ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>725.171ms<br/>0.725171ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>465.519ms<br/>0.465519ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>956.377ms<br/>0.956377ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>283.12ms<br/>0.28312ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>169.24ms<br/>0.16924ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>307.464ms<br/>0.307464ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>279.266ms<br/>0.279266ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>169.58ms<br/>0.16958ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>308.413ms<br/>0.308413ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>725.453ms<br/>0.725453ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>466.925ms<br/>0.466925ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>952.26ms<br/>0.95226ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>282.626ms<br/>0.282626ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>167.719ms<br/>0.167719ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>303.848ms<br/>0.303848ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>281.084ms<br/>0.281084ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>168.693ms<br/>0.168693ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>303.77ms<br/>0.30377ms/req |
| rtree | 1 result:<br/>208.76ms ->  0.020876 ms/query<br/>10 results:<br/>243.918ms ->  0.0243918 ms/query | 1 result:<br/>209.796ms ->  0.0209796 ms/query<br/>10 results:<br/>246.411ms ->  0.0246411 ms/query |
<!-- BENCHMARK_RESULTS_END -->